add user db

+++ src/main/java/org/yorku/mymarketplace/MyMarketPlaceController.java
@@ -1,5 +1,7 @@
 package org.yorku.mymarketplace;
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.yorku.mymarketplace.entity.CatalogItemEntity;
@@ -13,10 +15,29 @@ import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 @RestController()
+@Slf4j
 public class MyMarketPlaceController {
 
     private final CatalogRepository catalogRepository;
     private final UserRepository userRepository;
+
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    public class UnAuthorizedException extends RuntimeException {
+
+        public UnAuthorizedException(String message) {
+            super(message);
+            log.warn(message);
+        }
+
+    }
+
+    private void checkIfAdmin(String token) {
+        Optional<UserEntity> byId = userRepository.findById(Long.valueOf(token));
+        Optional<UserEntity> userEntity = byId.filter(u -> "admin".equalsIgnoreCase(u.getRole()));
+        if (userEntity.isEmpty()) {
+            throw new UnAuthorizedException("Not authorized");
+        }
+    }
 
     public MyMarketPlaceController(CatalogRepository catalogRepository, UserRepository userRepository) {
         this.catalogRepository = catalogRepository;
@@ -36,20 +57,33 @@ public class MyMarketPlaceController {
         return all;
     }
 
+
     @PutMapping(value = "/api/catalog-item", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public String addNewCatalogItem(@RequestBody CatalogItemEntity entity) {
+    public String addNewCatalogItem(@RequestBody CatalogItemEntity entity,
+                                    @RequestHeader("authentication") String token) {
+        checkIfAdmin(token);
         catalogRepository.save(entity);
         return "{}";
     }
 
+    @GetMapping("/api/catalog-item/{id}")
+    public CatalogItemEntity getCatalogItem(@PathVariable("id") Long id) {
+
+        return catalogRepository.findById(id).get();
+    }
+
     @DeleteMapping("/api/catalog-item/{id}")
-    public String deleteNewCatalogItem(@PathVariable("id") Long id) {
+    public String deleteNewCatalogItem(@PathVariable("id") Long id,
+                                       @RequestHeader("authentication") String token) {
+        checkIfAdmin(token);
         catalogRepository.deleteById(id);
         return "{}";
     }
 
+
     @GetMapping("/api/users")
-    public List<UserEntity> users() {
+    public List<UserEntity> users(@RequestHeader("authentication") String token) {
+        checkIfAdmin(token);
         List<UserEntity> all = StreamSupport //
                 .stream(userRepository.findAll().spliterator(), false) //
                 .collect(Collectors.toList());
@@ -58,6 +92,7 @@ public class MyMarketPlaceController {
 
     @PostMapping(value = "/api/login", consumes = MediaType.APPLICATION_JSON_VALUE)
     public UserEntity login(@RequestBody UserEntity entity) {
+        log.info("login: {}", entity);
         Optional<UserEntity> byName = userRepository.findByName(entity.getName());
         UserEntity user = byName.orElseThrow(() -> new IllegalStateException("user could not be found"));
         if (!entity.getPassword().equals(user.getPassword())) {
@@ -67,15 +102,37 @@ public class MyMarketPlaceController {
         return user;
     }
 
-    @PutMapping(value = "/api/user", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public String addUser(@RequestBody UserEntity entity) {
-        userRepository.save(entity);
-        return "{}";
+    @PutMapping(value = "/api/add-user", consumes = MediaType.APPLICATION_JSON_VALUE)
+    public UserEntity createNewAppUser(@RequestBody UserEntity entity,
+                                       @RequestHeader("authentication") String token) {
+        log.info("/api/add-user, {}", entity);
+//      TODO: add security check
+//        if (entity.getId() != null) {
+//            checkIfAdmin(token);
+//        }
+        List<UserEntity> byName = userRepository.checkNames(entity.getName());
+        if (entity.getId() == null && byName.size() > 0) {
+            throw new IllegalArgumentException("user already exists");
+        }
+        UserEntity saved = userRepository.save(entity);
+        entity.setId(saved.getId());
+        return entity;
     }
 
     @DeleteMapping("/api/user/{id}")
-    public String deleteUser(@PathVariable("id") Long id) {
+    public String deleteUser(@PathVariable("id") Long id,
+                             @RequestHeader("authentication") String token) {
+        checkIfAdmin(token);
         userRepository.deleteById(id);
         return "{}";
+    }
+
+    @GetMapping("/api/user/{id}")
+    public UserEntity selectUser(@PathVariable("id") Long id,
+                                 @RequestHeader("authentication") String token) {
+        //checkIfAdmin(token);
+        // TODO, check for self + admin
+
+        return userRepository.findById(id).get();
     }
 }
